Fix documentation Sersic1D and Sersic2D
The documentation for both the Sersic1D and Sersic2D gives the wrong model formula (missing a set of brackets).  As can be seen in the reference and the code, the factor -b_n should be multiplied with the whole (r/r_e)^(1/n) -1 term and not just (r/r_e)^(1/n).  I have fixed this by changing the square brackets to curly brackets and putting the (r/r_e)^(1/n) -1 term in square brackets as it is done in the reference.